# src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Marvel API",
        "description": "The Marvel Comics API allows developers to access information on Marvel characters",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:8080",
    "basePath": "/",
    "tags": [
        {
            "name": "characters",
            "description": "Fetches Marvel characters endpoints"
        }
    ],
    "paths": {
        "/characters": {
            "get": {
                "tags": [
                    "characters"
                ],
                "summary": "Fetches all Marvel's characters id",
                "description": "Fetches lists of id of Marvel comic characters",
                "responses": {
                    "200": {
                        "description": "list of ids in integer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "example": "[1011334,1017100,1009144,1010699,1009146,1016823,...]"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/characters/{id}": {
            "get": {
                "tags": [
                    "characters"
                ],
                "summary": "Fetches a single character by id",
                "description": "Fetches a single character by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Fetches a single character by id",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "character data",
                        "schema": {
                            "$ref": "#/definitions/Character"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Character": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Character ID",
                    "example": "1009146"
                },
                "name": {
                    "type": "string",
                    "description": "Character Name",
                    "example": "Abomination (Emil Blonsky)"
                },
                "description": {
                    "type": "string",
                    "description": "Character Description",
                    "example": "Formerly known as Emil Blonsky, a spy of Soviet Yugoslavian origin working for the KGB, the Abomination gained his powers after receiving a dose of gamma radiation similar to that which transformed Bruce Banner into the incredible Hulk."
                }
            }
        }
    },
    "schemes": [
        "HTTP"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ]
}